feat: add eslint
