+++ .eslintignore
@@ -1,0 +1,3 @@
+docs/
+dist/
+node_modules
