+++ .eslintrc.json
@@ -1,0 +1,35 @@
+{
+    "env": {
+        "browser": true,
+        "es2021": true
+    },
+    "extends": [
+        "plugin:react/recommended",
+        "plugin:react-hooks/recommended",
+        "nextjs",
+        "airbnb",
+        "prettier"
+    ],
+    "parserOptions": {
+        "ecmaFeatures": {
+            "jsx": true
+        },
+        "ecmaVersion": 12,
+        "sourceType": "module"
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "react/react-in-jsx-scope": "off",
+        "prettier/prettier": ["error", {
+           "endOfLine":"auto",
+           "singleQuote": true,
+           "jsxBracketSameLine": false
+         }]
+    },
+    "globals":{
+        "React": "writable",
+        "prop-types": "writable"
+    }
+}
