+++ components/nav.jsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import Link from 'next/link';
+
+const links = [
+  { href: 'https://zeit.co/now', label: 'ZEIT' },
+  { href: 'https://github.com/zeit/next.js', label: 'GitHub' },
+].map((link) => {
+  const newLink = { ...link };
+  newLink.key = `nav-link-${link.href}-${link.label}`;
+  return newLink;
+});
+
+const Nav = () => (
+  <nav>
+    <ul>
+      <li>
+        <Link href="/">
+          <a href="/">Home</a>
+        </Link>
+      </li>
+      {links.map(({ key, href, label }) => (
+        <li key={key}>
+          <a href={href}>{label}</a>
+        </li>
+      ))}
+    </ul>
+
+    <style jsx>
+      {`
+        nav {
+          text-align: center;
+        }
+        ul {
+          display: flex;
+          justify-content: space-between;
+        }
+        nav > ul {
+          padding: 4px 16px;
+        }
+        li {
+          display: flex;
+          padding: 6px 8px;
+        }
+        a {
+          color: #067df7;
+          text-decoration: none;
+          font-size: 13px;
+        }
+      `}
+    </style>
+  </nav>
+);
+
+export default Nav;
