+++ lib/gtag.js
@@ -1,17 +1,17 @@
-export const GA_TRACKING_ID = 'UA-83411453-5'
+export const GA_TRACKING_ID = 'UA-83411453-5';
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
-export const pageview = url => {
+export const pageview = (url) => {
   window.gtag('config', GA_TRACKING_ID, {
-    page_path: url
-  })
-}
+    page_path: url,
+  });
+};
 
 // https://developers.google.com/analytics/devguides/collection/gtagjs/events
 export const event = ({ action, category, label, value }) => {
   window.gtag('event', action, {
     event_category: category,
     event_label: label,
-    value: value
-  })
-}
+    value,
+  });
+};
