+++ next.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  exportPathMap: function () {
+  exportPathMap() {
     return {
       '/': { page: '/' },
     };
