+++ package.json
@@ -6,20 +6,35 @@
     "build": "next build",
     "dev": "cross-env NODE_ENV=development next dev",
     "export": "npm run build && next export -o docs/ && touch docs/.nojekyll && touch docs/CNAME && echo calicreators.com>> docs/CNAME",
-    "start": "next start"
+    "start": "next start",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
   "dependencies": {
     "isomorphic-unfetch": "^3.1.0",
     "moment": "^2.29.1",
     "next": "9.5.5",
-    "react": "16.13.1",
-    "react-dom": "16.13.1"
+    "prop-types": "^15.7.2",
+    "react": "16.14.0",
+    "react-dom": "16.14.0"
   },
   "devDependencies": {
     "@types/node": "^14.11.8",
     "@types/react": "^16.9.52",
+    "babel-eslint": "^10.1.0",
     "cross-env": "^7.0.2",
+    "eslint": "^7.11.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-nextjs": "^1.0.6",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-html": "^6.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.21.4",
+    "eslint-plugin-react-hooks": "^4.1.2",
     "husky": "^4.3.0",
+    "prettier": "^2.1.2",
     "typescript": "^4.0.3"
   }
 }
