+++ pages/[group]/extralife.jsx
@@ -1,15 +1,39 @@
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import fetch from 'isomorphic-unfetch';
-import Head from 'next/head';
+// import Head from 'next/head';
 import TeamMemberCards from '../../components/molecules/teamMemberCards';
 import Header from '../../components/atoms/header';
 import LoadingIcon from '../../components/atoms/loading-icon';
 import ProgressBar from '../../components/atoms/progress-bar';
 import StreamerSchedule from '../../components/organisms/streamer-schedule';
+
 const data = require('./data.json');
 
 const API_BASE = 'https://www.extra-life.org/api';
+
+function sortParticipants(participants) {
+  participants
+    .sort((a, b) => {
+      return a.displayName.localeCompare(b.displayName);
+    })
+    .sort((a, b) => {
+      if (a.streamIsLive) return -1;
+      if (b.streamIsLive) return 1;
+      return 0;
+    })
+    .map((member) => {
+      if (member?.links?.stream) {
+        const { stream } = member.links;
+        const regex = stream.match(/channel=(.*)/);
+        if (regex?.[1]) {
+          const newMember = { ...member, twitchUsername: regex[1] };
+          return newMember;
+        }
+      }
+      return member;
+    });
+}
 
 const ExtraLifeTeam = () => {
   const router = useRouter();
@@ -21,13 +45,13 @@ const ExtraLifeTeam = () => {
   useEffect(() => {
     async function fetchTeam() {
       const res = await fetch(`${API_BASE}/teams/${groupData.id}`);
-      const team = await res.json();
-      return team;
+      const fetchedTeam = await res.json();
+      return fetchedTeam;
     }
     async function fetchTeamMembers() {
       const res = await fetch(`${API_BASE}/teams/${groupData.id}/participants`);
-      const teamMembers = await res.json();
-      return teamMembers;
+      const fetchedTeamMembers = await res.json();
+      return fetchedTeamMembers;
     }
     async function getData() {
       const results = await Promise.all([fetchTeam(), fetchTeamMembers()]);
@@ -35,33 +59,14 @@ const ExtraLifeTeam = () => {
       // TODO REMOVE THIS
       results[1][11].streamIsLive = true;
       //!
-      const participants = results[1]
-        .sort((a, b) => {
-          return a.displayName.localeCompare(b.displayName);
-        })
-        .sort((a, b) => {
-          if (a.streamIsLive) return -1;
-          if (b.streamIsLive) return 1;
-          return 0;
-        })
-        .map((member) => {
-          if (member?.links?.stream) {
-            const stream = member.links.stream;
-            const regex = stream.match(/channel=(.*)/);
-            if (regex?.[1]) {
-              member.twitchUsername = regex[1];
-              return member;
-            }
-          }
-          return member;
-        });
+      const participants = sortParticipants(results[1]);
 
-      const team = {
+      const newTeam = {
         ...results[0],
         participants,
       };
-      console.log(team);
-      setTeam(team);
+      console.log(newTeam);
+      setTeam(newTeam);
     }
     if (groupData?.id) getData();
   }, [groupData]);
@@ -70,9 +75,9 @@ const ExtraLifeTeam = () => {
 
   console.log(groupData.schedule);
   return (
-    <div className={'page'}>
+    <div className="page">
       <Header title={team.name} />
-      <a href={team.links.page} className={'teamLink'}>
+      <a href={team.links.page} className="teamLink">
         Join Team
       </a>
       <ProgressBar
@@ -80,7 +85,7 @@ const ExtraLifeTeam = () => {
         goal={team.fundraisingGoal}
         progressText="Raised"
         goalText="Goal"
-        isMoney={true}
+        isMoney
         width={80}
       />
       <TeamMemberCards teamMembers={team.participants} />
@@ -90,16 +95,18 @@ const ExtraLifeTeam = () => {
           teamMembers={team.participants}
         />
       )}
-      <style jsx>{`
-        .page {
-          width: 100%;
-          text-align: center;
-        }
-        .teamLink {
-          color: #fff;
-          font-size: 24px;
-        }
-      `}</style>
+      <style jsx>
+        {`
+          .page {
+            width: 100%;
+            text-align: center;
+          }
+          .teamLink {
+            color: #fff;
+            font-size: 24px;
+          }
+        `}
+      </style>
     </div>
   );
 };
